Split AccessLogValve in an AbstractAccessLogValve with the formatting logic and an AccessLogValve with the logic to write logs into files.
